requirements.txt
aiohttp==3.6.3
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
better-profanity==0.7.0
braceexpand==0.1.7
cached-property==1.5.2
catalogue==1.0.0
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
discord-ext-menus==1.1
discord.py @ git+https://github.com/Rapptz/discord.py@58ca9e99325ac2678a51cd63afd7ae917f14bc8d
DiscordUtils==1.3.4
idna==3.2
import-expression==1.1.4
inflection==0.5.1
jishaku==2.2.0
keyboard==0.13.5
lavalink==3.1.4
lxml==4.6.3
more-itertools==8.8.0
multidict==4.7.6
murmurhash==1.0.5
numpy==1.21.1
ordered-set==3.1.1
ordered-set-stubs==0.1.3
plac==1.1.3
preshed==3.0.5
profanity-filter==1.3.3
profanityfilter==2.0.6
psutil==5.8.0
pycparser==2.20
pydantic==1.8.2
pyfiglet==0.8.post1
PyNaCl==1.4.0
pyperclip==1.8.2
pytube==11.0.0
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.15.100
six==1.16.0
tomlkit==0.5.11
typing-extensions==3.10.0.0
unicode==2.8
Unidecode==1.2.0
urllib3==1.26.6
wavelink==1.0.0b30
yarl==1.5.1